runtime (gc_blocks.go): use a linked list to scan marked objects
This adds a header to the start of each object which is used to track whether the object has been scanned yet.
This header now includes the layout information when using a precise collector.

It also fixes a bug where the layout field is updated outside of the GC lock.
This can cause the GC to read uninitialized memory and segfault if it runs immediately after the lock is released.