runtime (gc_blocks.go): simplify heap scanning
Previously, the GC would create a "scanner" and then use it to check if each position in a heap object is a pointer.
This moves the the iteration logic into the layout code, instead looping over valid positions.
It also replaces the custom marking logic with the existing markRoot function.