runtime (gc_blocks.go): optimize sweep
This changes the sweep pass to operate on state bytes instead of individual blocks.
Common cases (all tails or trivially unchanged state bytes) are handled specially with fast paths.
Otherwise this uses a simple branchless fallback to process the state byte updates.

In order to make this work, the bits for each block state were deinterleaved.
This allows us to efficiently get a mask of all blocks of a given type.